feat(query): Allow custom sttpbackend in RemoteExec
