+++ coordinator/src/main/scala/filodb.coordinator/queryplanner/MultiPartitionPlanner.scala
@@ -24,6 +24,8 @@ class MultiPartitionPlanner(partitionLocationProvider: PartitionLocationProvider
 
   import net.ceedubs.ficus.Ficus._
 
+  val sttpBackend = PlannerUtil.getSttpBackend(queryConfig.routingConfig)
+
   val remoteHttpTimeoutMs: Long =
     queryConfig.routingConfig.config.as[Option[Long]]("remote.http.timeout").getOrElse(60000)
 
@@ -118,7 +120,7 @@ class MultiPartitionPlanner(partitionLocationProvider: PartitionLocationProvider
         else {
           val httpEndpoint = p.endPoint + queryParams.remoteQueryPath.getOrElse("")
           PromQlRemoteExec(httpEndpoint, remoteHttpTimeoutMs, generateRemoteExecParams(qContext, startMs, endMs),
-            InProcessPlanDispatcher, dataset.ref)
+            InProcessPlanDispatcher, dataset.ref, sttpBackend)
         }
       }
       if (execPlans.size == 1) execPlans.head
@@ -144,7 +146,7 @@ class MultiPartitionPlanner(partitionLocationProvider: PartitionLocationProvider
         else {
           val httpEndpoint = partitions.head.endPoint + queryParams.remoteQueryPath.getOrElse("")
           PromQlRemoteExec(httpEndpoint, remoteHttpTimeoutMs, generateRemoteExecParams(qContext,
-            queryParams.startSecs * 1000, queryParams.endSecs * 1000), InProcessPlanDispatcher, dataset.ref
+            queryParams.startSecs * 1000, queryParams.endSecs * 1000), InProcessPlanDispatcher, dataset.ref, sttpBackend
             )
         }
       }
@@ -191,6 +193,6 @@ class MultiPartitionPlanner(partitionLocationProvider: PartitionLocationProvider
     val httpEndpoint = partitionAssignment.endPoint + finalQueryContext.origQueryParams.asInstanceOf[PromQlQueryParams].
       remoteQueryPath.getOrElse("")
     MetadataRemoteExec(httpEndpoint, remoteHttpTimeoutMs,
-      urlParams, finalQueryContext, InProcessPlanDispatcher, dataset.ref)
+      urlParams, finalQueryContext, InProcessPlanDispatcher, dataset.ref, sttpBackend)
   }
 }
