+++ coordinator/src/test/scala/filodb.coordinator/queryplanner/MultiPartitionPlannerSpec.scala
@@ -26,7 +26,8 @@ class MultiPartitionPlannerSpec extends AnyFunSpec with Matchers {
   private val dataset = MetricsTestData.timeseriesDataset
   private val schemas = Schemas(dataset.schema)
 
-  private val routingConfigString = "routing {\n  remote {\n    http {\n      timeout = 10000\n    }\n  }\n}"
+  private val routingConfigString = "routing {\n  remote {\n    http {\n      timeout = 10000\n" +
+    "      client.sttp-backend.factory = \"filodb.query.exec.AsyncHttpClientBackendFactory\"\n    }\n  }\n}"
   private val routingConfig = ConfigFactory.parseString(routingConfigString)
 
   private val config = ConfigFactory.load("application_test.conf")
