+++ query/src/main/scala/filodb/query/exec/PromQlRemoteExec.scala
@@ -1,96 +1,26 @@
 package filodb.query.exec
 
-import java.util.concurrent.TimeUnit
-
 import scala.concurrent.Future
-import scala.concurrent.duration._
-import scala.sys.ShutdownHookThread
 
-import com.softwaremill.sttp.asynchttpclient.future.AsyncHttpClientFutureBackend
-import com.softwaremill.sttp.circe._
-import com.typesafe.scalalogging.StrictLogging
+import com.softwaremill.sttp.SttpBackend
 import kamon.Kamon
 import kamon.trace.Span
-import monix.eval.Task
 import monix.execution.Scheduler
 
 import filodb.core.DatasetRef
 import filodb.core.metadata.Column.ColumnType
 import filodb.core.query._
-import filodb.core.store.ChunkSource
 import filodb.memory.format.ZeroCopyUTF8String._
 import filodb.memory.format.vectors.{CustomBuckets, MutableHistogram}
 import filodb.query._
 import filodb.query.AggregationOperator.Avg
 
-trait RemoteExec extends LeafExecPlan with StrictLogging {
-
-  def queryEndpoint: String
-
-  def requestTimeoutMs: Long
-
-  def urlParams: Map[String, Any]
-  def promQlQueryParams: PromQlQueryParams = queryContext.origQueryParams.asInstanceOf[PromQlQueryParams]
-
-  def args: String = s"${promQlQueryParams.toString}, ${queryContext.plannerParams}, queryEndpoint=$queryEndpoint, " +
-    s"requestTimeoutMs=$requestTimeoutMs"
-
-  def limit: Int = ???
-
-  /**
-   * Since execute is already overrided here, doExecute() can be empty.
-   */
-  def doExecute(source: ChunkSource,
-                querySession: QuerySession)
-               (implicit sched: Scheduler): ExecResult = ???
-
-  override def execute(source: ChunkSource,
-                       querySession: QuerySession)
-                      (implicit sched: Scheduler): Task[QueryResponse] = {
-    val execPlan2Span = Kamon.spanBuilder(s"execute-${getClass.getSimpleName}")
-      .asChildOf(Kamon.currentSpan())
-      .tag("query-id", queryContext.queryId)
-      .start()
-
-    if (queryEndpoint == null) {
-      throw new BadQueryException("Remote Query endpoint can not be null in RemoteExec.")
-    }
-
-    // Please note that the following needs to be wrapped inside `runWithSpan` so that the context will be propagated
-    // across threads. Note that task/observable will not run on the thread where span is present since
-    // kamon uses thread-locals.
-    Kamon.runWithSpan(execPlan2Span, true) {
-      Task.fromFuture(sendHttpRequest(execPlan2Span, requestTimeoutMs))
-    }
-  }
-
-  def sendHttpRequest(execPlan2Span: Span, httpTimeoutMs: Long)
-                     (implicit sched: Scheduler): Future[QueryResponse]
-
-  def getUrlParams(): Map[String, Any] = {
-    var finalUrlParams = urlParams ++
-      Map("start" -> promQlQueryParams.startSecs,
-        "end" -> promQlQueryParams.endSecs,
-        "time" -> promQlQueryParams.endSecs,
-        "step" -> promQlQueryParams.stepSecs,
-        "processFailure" -> queryContext.plannerParams.processFailure,
-        "processMultiPartition" -> queryContext.plannerParams.processMultiPartition,
-        "histogramMap" -> "true",
-        "skipAggregatePresent" -> queryContext.plannerParams.skipAggregatePresent,
-        "verbose" -> promQlQueryParams.verbose)
-    if (queryContext.plannerParams.spread.isDefined) finalUrlParams = finalUrlParams + ("spread" -> queryContext.
-      plannerParams.spread.get)
-    logger.debug("URLParams for RemoteExec:" + finalUrlParams)
-    finalUrlParams
-  }
-
-}
-
 case class PromQlRemoteExec(queryEndpoint: String,
                             requestTimeoutMs: Long,
                             queryContext: QueryContext,
                             dispatcher: PlanDispatcher,
-                            dataset: DatasetRef) extends RemoteExec {
+                            dataset: DatasetRef,
+                            sttpBackend: SttpBackend[Future, Nothing] = DefaultSttpBackend()) extends RemoteExec {
   private val defaultColumns = Seq(ColumnInfo("timestamp", ColumnType.TimestampColumn),
     ColumnInfo("value", ColumnType.DoubleColumn))
 
@@ -114,7 +44,7 @@ case class PromQlRemoteExec(queryEndpoint: String,
 
   override def sendHttpRequest(execPlan2Span: Span, httpTimeoutMs: Long)
                               (implicit sched: Scheduler): Future[QueryResponse] = {
-    PromRemoteExec.httpGet(queryEndpoint, requestTimeoutMs, queryContext.submitTime, getUrlParams())
+    RemoteHttpClient(sttpBackend).httpGet(queryEndpoint, requestTimeoutMs, queryContext.submitTime, getUrlParams())
       .map { response =>
         response.unsafeBody match {
           case Left(error) => QueryError(queryContext.queryId, error.error)
@@ -252,51 +182,3 @@ case class PromQlRemoteExec(queryEndpoint: String,
   }
 
 }
-
-object PromRemoteExec extends StrictLogging {
-
-  import com.softwaremill.sttp._
-  import io.circe.generic.auto._
-
-  // DO NOT REMOVE PromCirceSupport import below assuming it is unused - Intellij removes it in auto-imports :( .
-  // Needed to override Sampl case class Encoder.
-  import PromCirceSupport._
-  implicit val backend = AsyncHttpClientFutureBackend()
-
-  ShutdownHookThread(shutdown())
-
-  def httpGet(httpEndpoint: String, httpTimeoutMs: Long, submitTime: Long, urlParams: Map[String, Any])
-  (implicit scheduler: Scheduler):
-  Future[Response[scala.Either[DeserializationError[io.circe.Error], SuccessResponse]]] = {
-    val queryTimeElapsed = System.currentTimeMillis() - submitTime
-    val readTimeout = FiniteDuration(httpTimeoutMs - queryTimeElapsed, TimeUnit.MILLISECONDS)
-    val url = uri"$httpEndpoint?$urlParams"
-    logger.debug("promQlExec url={}", url)
-    sttp
-      .get(url)
-      .readTimeout(readTimeout)
-      .response(asJson[SuccessResponse])
-      .send()
-  }
-
-  def httpMetadataGet(httpEndpoint: String, httpTimeoutMs: Long, submitTime: Long, urlParams: Map[String, Any])
-                     (implicit scheduler: Scheduler):
-  Future[Response[scala.Either[DeserializationError[io.circe.Error], MetadataSuccessResponse]]] = {
-    val queryTimeElapsed = System.currentTimeMillis() - submitTime
-    val readTimeout = FiniteDuration(httpTimeoutMs - queryTimeElapsed, TimeUnit.MILLISECONDS)
-    val url = uri"$httpEndpoint?$urlParams"
-    logger.debug("promMetadataExec url={}", url)
-    sttp
-      .get(url)
-      .readTimeout(readTimeout)
-      .response(asJson[MetadataSuccessResponse])
-      .send()
-  }
-
-  def shutdown(): Unit =
-  {
-    logger.info("Shutting PromQlExec http")
-    backend.close()
-  }
-}
-
